refactor: Extract propsal details into a seprate component for reusability
