+++ lib/components/proposal/index.ts
@@ -1,4 +1,5 @@
 import { ProposalCard, ProposalCardSkeleton } from "./ProposalCard";
 import ProposalContentSummary from "./ProposalSummaryCard";
+import ProposalDetails from "./ProposalDetails";
 
-export { ProposalCard, ProposalCardSkeleton, ProposalContentSummary };
+export { ProposalCard, ProposalCardSkeleton, ProposalContentSummary, ProposalDetails };
